refactor: move FakeWalletGen into integrationtests
