+++ fedimint-testing/Cargo.toml
@@ -18,7 +18,6 @@ async-trait = "*"
 bitcoin = "0.29.2"
 bitcoincore-rpc = "0.16.0"
 clap = { version = "4.1.6", features = ["derive", "std", "help", "usage", "error-context", "suggestions" ], default-features = false }
-erased-serde = "0.3"
 fedimint-client-legacy = { path = "../fedimint-client-legacy" }
 fedimint-core  = { path = "../fedimint-core" }
 fedimint-client  = { path = "../fedimint-client" }
@@ -35,7 +34,6 @@ lightning-invoice = "0.21.0"
 tempfile = "3.4.0"
 secp256k1-zkp = { version = "0.7.0", features = [ "global-context", "bitcoin_hashes" ] }
 serde = { version = "1.0.149", features = [ "derive" ] }
-serde_json = "1.0.91"
 tracing ="0.1.37"
 rand = "0.8"
 tokio = { version = "1.26.0", features = ["full"] }
