+++ integrationtests/tests/fixtures/mod.rs
@@ -53,7 +53,6 @@ use fedimint_server::net::connect::mock::{MockNetwork, StreamReliability};
 use fedimint_server::net::connect::{parse_host_port, Connector, TlsTcpConnector};
 use fedimint_server::net::peers::{DelayCalculator, PeerConnector};
 use fedimint_server::{consensus, FedimintApiHandler, FedimintServer};
-use fedimint_testing::btc::bitcoind::FakeWalletGen;
 use fedimint_testing::btc::fixtures::FakeBitcoinTest;
 use fedimint_testing::btc::BitcoinTest;
 use fedimint_testing::ln::fixtures::FakeLightningTest;
@@ -66,6 +65,7 @@ use fedimint_wallet_server::{Wallet, WalletGen};
 use futures::executor::block_on;
 use futures::future::{join_all, select_all};
 use futures::{FutureExt, StreamExt};
+use gen::FakeWalletGen;
 use hbbft::honey_badger::Batch;
 use legacy::LegacyTestUser;
 use ln_gateway::actor::GatewayActor;
@@ -86,6 +86,7 @@ use crate::fixtures::user::{IGatewayClient, ILegacyTestClient};
 use crate::fixtures::utils::LnRpcAdapter;
 use crate::ConsensusItem;
 
+mod gen;
 mod legacy;
 mod real;
 pub mod user;
